Hotspot initial commit
Creates the framework for having "hotspots" - areas that are restricted to only a certain number of players per allegiance.

This currently handles checking number of players in same allegiance in the hotspot area at the time of teleporting in (this is assuming its an enclosed area that can't be accessed by just walking in and doesn't allow relogging.

Future commit will add logic to kick allegiance members if entering a landblock without exiting a portal